Initial support for 1.16's true-color components
1.16 supports using any color for text components, so this mod has to support it too.

Right now there's a bug where only the first text component's style is considered, so this is not yet ready for release. I'm getting there though. Probably a bug in the visitor implementation...

I'd like to implement an option to not use true color (for older terminals) but for now I'm just assuming true color is supported. Also that would require a configuration file which I'm not sure about yet.

This commit also adds a new command, /tellrawc, that logs text components to console. For testing but also for the lulz.